Refactor navigation logic
This commit includes the following changes:

- Rearranged and sorted the codebase for better organization
- Added a new navigation graph for the app gate, including OnBoardingFragment and FragmentLogin and FragmentSignUp
- Implemented navigation logic for the logout functionality
- Implemented logic to show the OnBoardingFragment if the user hasn't completed onBoarding

These changes improve the navigation flow of the app and introduce a new feature for onBoarding new users. The codebase has been cleaned and optimized for better readability and maintainability.

Issue: #<1>
The First step of building a Notify System that alert the user when ever the medicine time is up, <It needs more work>
(There is a Database system but still haven't used it)